requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
PyMuPDF
Pillow